script_analyzer/requirements.txt
streamlit
groq
python-dotenv
pypdf
matplotlib
vaderSentiment
pandas
